Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -17,8 +17,8 @@ int main() {
             if(b==0){
                 printf("error");
             }
-                else{
-                    printf("%.0f",(float)a/b);
+            else{
+                    printf("%.0f",a/b);
                 }
             break;
         default:
